refactor(orc8r): Ctraced servicer endpoint as external
